Feature/minor improvements

+++ Classes/ViewHelpers/AbstractContentElementViewHelper.php
@@ -46,10 +46,10 @@ abstract class AbstractContentElementViewHelper extends AbstractPDFViewHelper
     {
         parent::initializeArguments();
 
-        $this->registerArgument('posX', 'integer', '', false, null);
-        $this->registerArgument('posY', 'integer', '', false, null);
+        $this->registerArgument('posX', 'float', '', false, null);
+        $this->registerArgument('posY', 'float', '', false, null);
         $this->registerArgument('width', 'string', '', false, null);
-        $this->registerArgument('height', 'integer', '', false, null);
+        $this->registerArgument('height', 'float', '', false, null);
     }
 
     /**
